+++ fastapi_utils/inferring_router.py
@@ -1,16 +1,11 @@
-from typing import TYPE_CHECKING, Any, Callable, get_type_hints
+from __future__ import annotations
+
+import warnings
 
 from fastapi import APIRouter
 
+warnings.warn(
+    "InferringRouter is deprecated, as its functionality is now provided in fastapi.APIRouter", DeprecationWarning
+)
 
-class InferringRouter(APIRouter):
-    """
-    Overrides the route decorator logic to use the annotated return type as the `response_model` if unspecified.
-    """
-
-    if not TYPE_CHECKING:  # pragma: no branch
-
-        def add_api_route(self, path: str, endpoint: Callable[..., Any], **kwargs: Any) -> None:
-            if kwargs.get("response_model") is None:
-                kwargs["response_model"] = get_type_hints(endpoint).get("return")
-            return super().add_api_route(path, endpoint, **kwargs)
+InferringRouter = APIRouter