+++ tests/test_cbv.py
@@ -1,4 +1,6 @@
-from typing import Any, ClassVar
+from __future__ import annotations
+
+from typing import Any, ClassVar, Optional
 
 from fastapi import APIRouter, Depends, FastAPI
 from starlette.testclient import TestClient
@@ -70,7 +72,7 @@ def test_multiple_decorators() -> None:
         @router.get("/items/?")
         @router.get("/items/{item_path:path}")
         @router.get("/database/{item_path:path}")
-        def root(self, item_path: str = None, item_query: str = None) -> Any:
+        def root(self, item_path: Optional[str] = None, item_query: Optional[str] = None) -> Any:
             if item_path:
                 return {"item_path": item_path}
             if item_query: