+++ docs/release-notes.md
@@ -1,5 +1,13 @@
 ## Latest changes
 
+## 0.3.0
+
+* Move to ruff for linting, etc.
+* Update various dependencies
+* Stop supporting Python 3.6
+* Deprecate InferringRouter (as its functionality is now built into `fastapi.APIRouter`)
+* Resolve various deprecationwarnings introduced by sqlalchemy 1.4.
+
 ## 0.2.1
 
 * Fix bug with multiple decorators on same method